Angular Forms Directive (#88)
This change allows for the use of `ngModel` and `formControlName` when using the drop-downs with Angular.

There should be no breaking changes to previous functionality or usage for either Angular or Vanilla NativeScript.
Build scripts updated. Compatible with Angular AoT and Webpack.

Based on `nativescript-angular/value-accessors/selectedIndex-value-accessor.ts` and `nativescript-angular/nativescript-angular/forms.ts` with some changes to conform to project linting.